+++ python/light-curve/CHANGELOG.md
@@ -33,6 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0]
 
+### Added
+
+- `check: bool = True` keyword argument for `__call__` and `many` methods of feature classes. It coulb be used to check if input arrays are valid
+
 ### Changed
 
 - `gsl` is a default Cargo feature now, which means that GSL must be installed to build this package